Revert "Pruebas automaticas + arreglo de base"
