+++ BrandApp/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from .models import *
 
-# admin.site.register(User)
+admin.site.register(User)
 admin.site.register(Product)
 admin.site.register(Category)
 admin.site.register(Order)
