+++ BrandApp/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class BrandControlConfig(AppConfig):
+class BrandappConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'brand_control'
+    name = 'BrandApp'
