--- user_control/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from user_control.models import Users
-
-
-# Register your models here.
-admin.site.register(Users)
