--- user_control/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class UserControlConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'user_control'
