--- user_control/permissions.py
@@ -1,5 +1,0 @@
-from rest_framework.permissions import BasePermission
-
-class IsAdminUserCustom(BasePermission):
-    def has_permission(self, request, view):
-        return request.user.is_authenticated and request.user.roles == 'admin'
